ticket_notifcation/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ibm_watson
ibm_cloud_sdk_core
SoftLayer
requests